requirements.txt
apache-airflow==2.9.2
pyspark==3.5.1
